Transform Deadman to an SFC
